+++ fabric/network.py
@@ -6,14 +6,17 @@ from __future__ import with_statement
 
 from functools import wraps
 import getpass
+import os
 import re
 import threading
+import time
 import select
 import socket
 import sys
 
 from fabric.auth import get_password, set_password
 from fabric.utils import abort, handle_prompt_abort
+from fabric.exceptions import NetworkError
 
 try:
     import warnings
@@ -64,16 +67,26 @@ class HostConnectionCache(dict):
     two different connections to the same host being made. If no port is given,
     22 is assumed, so ``example.com`` is equivalent to ``example.com:22``.
     """
-    def __getitem__(self, key):
-        # Normalize given key (i.e. obtain username and port, if not given)
+    def connect(self, key):
+        """
+        Force a new connection to ``key`` host string.
+        """
         user, host, port = normalize(key)
-        # Recombine for use as a key.
-        real_key = join_host_strings(user, host, port)
-        # If not found, create new connection and store it
-        if real_key not in self:
-            self[real_key] = connect(user, host, port)
-        # Return the value either way
-        return dict.__getitem__(self, real_key)
+        key = normalize_to_string(key)
+        self[key] = connect(user, host, port)
+
+    def __getitem__(self, key):
+        """
+        Autoconnect + return connection object
+        """
+        key = normalize_to_string(key)
+        if key not in self:
+            self.connect(key)
+        return dict.__getitem__(self, key)
+
+    #
+    # Dict overrides that normalize input keys
+    #
 
     def __setitem__(self, key, value):
         return dict.__setitem__(self, normalize_to_string(key), value)
@@ -84,22 +97,100 @@ class HostConnectionCache(dict):
     def __contains__(self, key):
         return dict.__contains__(self, normalize_to_string(key))
 
+
+def ssh_config(host_string=None):
+    """
+    Return ssh configuration dict for current env.host_string host value.
+
+    Memoizes the loaded SSH config file, but not the specific per-host results.
+
+    This function performs the necessary "is SSH config enabled?" checks and
+    will simply return an empty dict if not. If SSH config *is* enabled and the
+    value of env.ssh_config_path is not a valid file, it will abort.
+
+    May give an explicit host string as ``host_string``.
+    """
+    from fabric.state import env
+    if not env.use_ssh_config:
+        return {}
+    if '_ssh_config' not in env:
+        try:
+            conf = ssh.SSHConfig()
+            path = os.path.expanduser(env.ssh_config_path)
+            with open(path) as fd:
+                conf.parse(fd)
+                env._ssh_config = conf
+        except IOError, e:
+            abort("Unable to load SSH config file '%s'" % path)
+    host = parse_host_string(host_string or env.host_string)['host']
+    return env._ssh_config.lookup(host)
+
+
+def key_filenames():
+    """
+    Returns list of SSH key filenames for the current env.host_string.
+
+    Takes into account ssh_config and env.key_filename, including normalization
+    to a list. Also performs ``os.path.expanduser`` expansion on any key
+    filenames.
+    """
+    from fabric.state import env
+    keys = env.key_filename
+    # For ease of use, coerce stringish key filename into list
+    if not isinstance(env.key_filename, (list, tuple)):
+        keys = [keys]
+    # Strip out any empty strings (such as the default value...meh)
+    keys = filter(bool, keys)
+    # Honor SSH config
+    # TODO: fix ssh so it correctly treats IdentityFile as a list
+    conf = ssh_config()
+    if 'identityfile' in conf:
+        keys.append(conf['identityfile'])
+    return map(os.path.expanduser, keys)
+
+
+def parse_host_string(host_string):
+    return host_regex.match(host_string).groupdict()
+
+
 def normalize(host_string, omit_port=False):
     """
     Normalizes a given host string, returning explicit host, user, port.
 
     If ``omit_port`` is given and is True, only the host and user are returned.
+
+    This function will process SSH config files if Fabric is configured to do
+    so, and will use them to fill in some default values or swap in hostname
+    aliases.
     """
     from fabric.state import env
     # Gracefully handle "empty" input by returning empty output
     if not host_string:
         return ('', '') if omit_port else ('', '', '')
-    # Get user, host and port separately
-    r = host_regex.match(host_string).groupdict()
-    # Add any necessary defaults in
-    user = r['user'] or env.get('user')
+    # Parse host string (need this early on to look up host-specific ssh_config
+    # values)
+    r = parse_host_string(host_string)
     host = r['host']
-    port = r['port'] or '22'
+    # Env values (using defaults if somehow earlier defaults were replaced with
+    # empty values)
+    user = env.user or env.local_user
+    port = env.port or env.default_port
+    # SSH config data
+    conf = ssh_config(host_string)
+    # Only use ssh_config values if the env value appears unmodified from
+    # the true defaults. If the user has tweaked them, that new value
+    # takes precedence.
+    if user == env.local_user and 'user' in conf:
+        user = conf['user']
+    if port == env.default_port and 'port' in conf:
+        port = conf['port']
+    # Also override host if needed
+    if 'hostname' in conf:
+        host = conf['hostname']
+    # Merge explicit user/port values with the env/ssh_config derived ones
+    # (Host is already done at this point.)
+    user = r['user'] or user
+    port = r['port'] or port
     if omit_port:
         return user, host
     return user, host, port
@@ -160,7 +251,7 @@ def connect(user, host, port):
     """
     Create and return a new SSHClient instance connected to given host.
     """
-    from state import env
+    from state import env, output
 
     #
     # Initialization
@@ -183,18 +274,20 @@ def connect(user, host, port):
     # Initialize loop variables
     connected = False
     password = get_password()
+    tries = 0
 
     # Loop until successful connect (keep prompting for new password)
     while not connected:
         # Attempt connection
         try:
+            tries += 1
             client.connect(
                 hostname=host,
                 port=int(port),
                 username=user,
                 password=password,
-                key_filename=env.key_filename,
-                timeout=10,
+                key_filename=key_filenames(),
+                timeout=env.timeout,
                 allow_agent=not env.no_agent,
                 look_for_keys=not env.no_keys
             )
@@ -208,10 +301,8 @@ def connect(user, host, port):
         # BadHostKeyException corresponds to key mismatch, i.e. what on the
         # command line results in the big banner error about man-in-the-middle
         # attacks.
-        except ssh.BadHostKeyException:
-            abort("Host key for %s did not match pre-existing key! Server's"
-                   " key was changed recently, or possible man-in-the-middle"
-                   "attack." % env.host)
+        except ssh.BadHostKeyException, e:
+            raise NetworkError("Host key for %s did not match pre-existing key! Server's key was changed recently, or possible man-in-the-middle attack." % env.host, e)
         # Prompt for new password to try on auth failure
         except (
             ssh.AuthenticationException,
@@ -225,7 +316,7 @@ def connect(user, host, port):
             # SSHException and there *was* a password -- it is probably
             # something non auth related, and should be sent upwards.
             if e.__class__ is ssh.SSHException and password:
-                abort(str(e))
+                raise NetworkError(str(e), e)
 
             # Otherwise, assume an auth exception, and prompt for new/better
             # password.
@@ -265,18 +356,41 @@ def connect(user, host, port):
             # Print a newline (in case user was sitting at prompt)
             print('')
             sys.exit(0)
-        # Handle timeouts
-        except socket.timeout:
-            abort('Timed out trying to connect to %s' % host)
         # Handle DNS error / name lookup failure
-        except socket.gaierror:
-            abort('Name lookup failed for %s' % host)
-        # Handle generic network-related errors
+        except socket.gaierror, e:
+            raise NetworkError('Name lookup failed for %s' % host, e)
+        # Handle timeouts and retries, including generic errors
         # NOTE: In 2.6, socket.error subclasses IOError
         except socket.error, e:
-            abort('Low level socket error connecting to host %s: %s' % (
-                host, e[1])
-            )
+            not_timeout = type(e) is not socket.timeout
+            giving_up = tries >= env.connection_attempts
+            # Baseline error msg for when debug is off
+            msg = "Timed out trying to connect to %s" % host
+            # Expanded for debug on
+            err = msg + " (attempt %s of %s)" % (tries, env.connection_attempts)
+            if giving_up:
+                err += ", giving up"
+            err += ")"
+            # Debuggin'
+            if output.debug:
+                print >>sys.stderr, err
+            # Having said our piece, try again
+            if not giving_up:
+                # Sleep if it wasn't a timeout, so we still get timeout-like
+                # behavior
+                if not_timeout:
+                    time.sleep(env.timeout)
+                continue
+            # Override eror msg if we were retrying other errors
+            if not_timeout:
+                msg = "Low level socket error connecting to host %s: %s" % (
+                    host, e[1]
+                )
+            # Here, all attempts failed. Tweak error msg to show # tries.
+            # TODO: find good humanization module, jeez
+            s = "s" if env.connection_attempts > 1 else ""
+            msg += " (tried %s time%s)" % (env.connection_attempts, s)
+            raise NetworkError(msg, e)
 
 
 def prompt_for_password(prompt=None, no_colon=False, stream=None):