+++ fabric/tasks.py
@@ -3,11 +3,18 @@ from __future__ import with_statement
 from functools import wraps
 
 from fabric import state
-from fabric.utils import abort
+from fabric.utils import abort, warn, error
 from fabric.network import to_dict, normalize_to_string
 from fabric.context_managers import settings
 from fabric.job_queue import JobQueue
-from fabric.task_utils import *
+from fabric.task_utils import crawl, merge, parse_kwargs
+from fabric.exceptions import NetworkError
+
+
+def _get_list(env):
+    def inner(key):
+        return env.get(key, [])
+    return inner
 
 
 class Task(object):
@@ -61,7 +68,7 @@ class Task(object):
         # from the CLI or from module-level code). This will be the empty list
         # if these have not been set -- which is fine, this method should
         # return an empty list if no hosts have been set anywhere.
-        env_vars = map(env.get, "hosts roles exclude_hosts".split())
+        env_vars = map(_get_list(env), "hosts roles exclude_hosts".split())
         env_vars.append(roledefs)
         return merge(*env_vars)
 
@@ -131,6 +138,54 @@ def _parallel_tasks(commands_to_run):
     ))
 
 
+def _execute(task, host, my_env, args, kwargs, jobs, queue, multiprocessing):
+    """
+    Primary single-host work body of execute()
+    """
+    # Log to stdout
+    if state.output.running and not hasattr(task, 'return_value'):
+        print("[%s] Executing task '%s'" % (host, my_env['command']))
+    # Create per-run env with connection settings
+    local_env = to_dict(host)
+    local_env.update(my_env)
+    state.env.update(local_env)
+    # Handle parallel execution
+    if queue is not None: # Since queue is only set for parallel
+        # Set a few more env flags for parallelism
+        state.env.parallel = True # triggers some extra aborts, etc
+        state.env.linewise = True # to mirror -P behavior
+        name = local_env['host_string']
+        # Wrap in another callable that:
+        # * nukes the connection cache to prevent shared-access problems
+        # * knows how to send the tasks' return value back over a Queue
+        # * captures exceptions raised by the task
+        def inner(args, kwargs, queue, name):
+            key = normalize_to_string(state.env.host_string)
+            state.connections.pop(key, "")
+            try:
+                result = task.run(*args, **kwargs)
+            except BaseException, e: # We really do want to capture everything
+                result = e
+            queue.put({'name': name, 'result': result})
+
+        # Stuff into Process wrapper
+        kwarg_dict = {
+            'args': args,
+            'kwargs': kwargs,
+            'queue': queue,
+            'name': name
+        }
+        p = multiprocessing.Process(target=inner, kwargs=kwarg_dict)
+        # Name/id is host string
+        p.name = name
+        # Add to queue
+        jobs.append(p)
+    # Handle serial execution
+    else:
+        return task.run(*args, **kwargs)
+
+def _is_task(task):
+    return isinstance(task, Task)
 
 def execute(task, *args, **kwargs):
     """
@@ -155,15 +210,30 @@ def execute(task, *args, **kwargs):
     ``task`` when it is called, so ``execute(mytask, 'arg1', kwarg1='value')``
     will (once per host) invoke ``mytask('arg1', kwarg1='value')``.
 
+    This function returns a dictionary mapping host strings to the given task's
+    return value for that host's execution run. For example, ``execute(foo,
+    hosts=['a', 'b'])`` might return ``{'a': None, 'b': 'bar'}`` if ``foo``
+    returned nothing on host `a` but returned ``'bar'`` on host `b`.
+
+    In situations where a task execution fails for a given host but overall
+    progress does not abort (such as when :ref:`env.skip_bad_hosts
+    <skip-bad-hosts>` is True) the return value for that host will be the error
+    object or message.
+
     .. seealso::
         :ref:`The execute usage docs <execute>`, for an expanded explanation
         and some examples.
 
     .. versionadded:: 1.3
+    .. versionchanged:: 1.4
+        Added the return value mapping; previously this function had no defined
+        return value.
     """
     my_env = {}
+    results = {}
     # Obtain task
-    if not callable(task):
+    is_callable = callable(task)
+    if not (is_callable or _is_task(task)):
         # Assume string, set env.command to it
         my_env['command'] = task
         task = crawl(task, state.commands)
@@ -173,79 +243,75 @@ def execute(task, *args, **kwargs):
     else:
         dunder_name = getattr(task, '__name__', None)
         my_env['command'] = getattr(task, 'name', dunder_name)
-    # Normalize to Task instance
-    if not hasattr(task, 'run'):
+    # Normalize to Task instance if we ended up with a regular callable
+    if not _is_task(task):
         task = WrappedCallableTask(task)
     # Filter out hosts/roles kwargs
     new_kwargs, hosts, roles, exclude_hosts = parse_kwargs(kwargs)
     # Set up host list
     my_env['all_hosts'] = task.get_hosts(hosts, roles, exclude_hosts, state.env)
 
+    parallel = requires_parallel(task)
+    if parallel:
+        # Import multiprocessing if needed, erroring out usefully
+        # if it can't.
+        try:
+            import multiprocessing
+        except ImportError:
+            import traceback
+            tb = traceback.format_exc()
+            abort(tb + """
+    At least one task needs to be run in parallel, but the
+    multiprocessing module cannot be imported (see above
+    traceback.) Please make sure the module is installed
+    or that the above ImportError is fixed.""")
+    else:
+        multiprocessing = None
+
     # Get pool size for this task
     pool_size = task.get_pool_size(my_env['all_hosts'], state.env.pool_size)
     # Set up job queue in case parallel is needed
-    jobs = JobQueue(pool_size)
+    queue = multiprocessing.Queue() if parallel else None
+    jobs = JobQueue(pool_size, queue)
     if state.output.debug:
         jobs._debug = True
 
     # Call on host list
     if my_env['all_hosts']:
+        # Attempt to cycle on hosts, skipping if needed
         for host in my_env['all_hosts']:
-            # Log to stdout
-            if state.output.running and not hasattr(task, 'return_value'):
-                print("[%s] Executing task '%s'" % (host, my_env['command']))
-            # Create per-run env with connection settings
-            local_env = to_dict(host)
-            local_env.update(my_env)
-            state.env.update(local_env)
-            # Handle parallel execution
-            if requires_parallel(task):
-                # Set a few more env flags for parallelism
-                state.env.parallel = True # triggers some extra aborts, etc
-                state.env.linewise = True # to mirror -P behavior
-                # Import multiprocessing if needed, erroring out usefully
-                # if it can't.
-                try:
-                    import multiprocessing
-                except ImportError:
-                    import traceback
-                    tb = traceback.format_exc()
-                    abort(tb + """
-At least one task needs to be run in parallel, but the
-multiprocessing module cannot be imported (see above
-traceback.) Please make sure the module is installed
-or that the above ImportError is fixed.""")
-
-                # Wrap in another callable that nukes the child's cached
-                # connection object, if needed, to prevent shared-socket
-                # problems.
-                def inner(*args, **kwargs):
-                    key = normalize_to_string(state.env.host_string)
-                    state.connections.pop(key, "")
-                    task.run(*args, **kwargs)
-                # Stuff into Process wrapper
-                p = multiprocessing.Process(target=inner, args=args,
-                    kwargs=new_kwargs)
-                # Name/id is host string
-                p.name = local_env['host_string']
-                # Add to queue
-                jobs.append(p)
-            # Handle serial execution
-            else:
-                task.run(*args, **new_kwargs)
+            try:
+                results[host] = _execute(
+                    task, host, my_env, args, new_kwargs, jobs, queue,
+                    multiprocessing
+                )
+            except NetworkError, e:
+                results[host] = e
+                # Backwards compat test re: whether to use an exception or
+                # abort
+                if not state.env.use_exceptions_for['network']:
+                    func = warn if state.env.skip_bad_hosts else abort
+                    error(e.message, func=func, exception=e.wrapped)
+                else:
+                    raise
 
         # If running in parallel, block until job queue is emptied
         if jobs:
+            err = "One or more hosts failed while executing task '%s'" % (
+                my_env['command']
+            )
             jobs.close()
-            exitcodes = jobs.run()
             # Abort if any children did not exit cleanly (fail-fast).
             # This prevents Fabric from continuing on to any other tasks.
-            if any([x != 0 for x in exitcodes]):
-                abort("One or more hosts failed while executing task '%s'" % (
-                    my_env['command']
-                ))
+            # Otherwise, pull in results from the child run.
+            for name, d in jobs.run().iteritems():
+                if d['exit_code'] != 0:
+                    abort(err)
+                results[name] = d['results']
 
     # Or just run once for local-only
     else:
         state.env.update(my_env)
-        task.run(*args, **new_kwargs)
+        results['<local-only>'] = task.run(*args, **new_kwargs)
+    # Return what we can from the inner task executions
+    return results